requirements-lite.txt
beautifulsoup4==4.10.0
lxml==4.6.4
numpy==1.21.4
Pillow==8.4.0
requests==2.26.0
retrying==1.3.3
